Add idempotency to clean command
If the VM is already shut off we should proceed to undefine it and
remove the disk without erroring. If the destroy fails for some other
reason we should indeed fail there, so only check that the state is not
'shut off' rather than that it is 'running'.

+++ scripts/clean.sh
@@ -9,7 +9,9 @@ clean_vm() {
         force=true
     fi
     local disk=$(virsh domblklist $vmname | awk 'NR==3 {print  $2}')
-    virsh destroy $vmname
+    if ! virsh domstate $vmname | grep 'shut off' ; then
+        virsh destroy $vmname
+    fi
     virsh undefine $vmname
     if [ "$force" == "true" ] ; then
         rm $disk -f
